Fix the javadoc and the formatting of the base-classes for objects with native references
